Fixed crash bug when app is minimized

+++ Source/Framework/Scripting/TypeRepresentation.cs
@@ -17,7 +17,7 @@ namespace Pegasus.Framework.Scripting
 		{
 			{ typeof(bool), o => o.ToString().ToLower() },
 			{ typeof(float), o => ((float)o).ToString("F") },
-			{ typeof(double), o => ((float)o).ToString("F") },
+			{ typeof(double), o => ((double)o).ToString("F") },
 			{ typeof(string), o => String.Format("\"{0}\"", o.ToString().Replace("\"", "\\\"")) },
 			{ typeof(InputTrigger), o => String.Format("[{0}]", o) }
 		};
